refactor: simplify the main app
